pyproject.toml
[project]
name = "customllm"
version = "0.1.0"
description = ""
authors = [
    {name = "Ivanxiedata",email = "ivanxiedata@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "streamlit (==1.42.0)",
    "langchain-core (==0.3.34)",
    "langchain-community (==0.3.17)",
    "langchain-ollama (==0.2.3)",
    "pdfplumber (==0.11.5)",
    "ollama (>=0.4.7,<0.5.0)",
    "torchvision (>=0.21.0,<0.22.0)",
    "fpdf (>=1.7.2,<2.0.0)",
    "fpdf2 (>=2.8.2,<3.0.0)",
    "reportlab (>=4.3.1,<5.0.0)",
    "chainlit (>=2.2.0,<3.0.0)",
    "chromadb (>=0.6.3,<0.7.0)",
    "groq (>=0.18.0,<0.19.0)",
    "langchain-groq (>=0.2.4,<0.3.0)",
    "pypdf2 (>=3.0.1,<4.0.0)",
    "faiss-cpu (>=1.10.0,<2.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "pytesseract (>=0.3.13,<0.4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
